+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## Next
 
+- Added Consul Module
+- Added Vault Module
 - Added Traefik Module